pyproject.toml
[project]
name = "blackpressarchive"
version = "0.1.0"
description = "Dynamic timeline browser for historical Black press newspapers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "pyyaml>=6.0",
]
